+++ test/conftest.py
@@ -1,5 +1,11 @@
 import os
 import contextlib
+from os.path import abspath, join, dirname
+from .tools import run_asv_with_conf
+import pytest
+from asv import config
+import shutil
+from .tools import locked_cache_dir
 
 
 def pytest_addoption(parser):
@@ -42,3 +48,27 @@ def _monkeypatch_conda_lock(config):
 
     path = config.cache.makedir('conda-lock') / 'lock'
     asv.plugins.conda._conda_lock = _conda_lock
+
+
+@pytest.fixture(scope="session")
+def example_results(request):
+    with locked_cache_dir(request.config, "example-results") as cache_dir:
+        src = abspath(join(dirname(__file__), 'example_results'))
+        dst = abspath(join(cache_dir, 'results'))
+
+        if os.path.isdir(dst):
+            return dst
+
+        shutil.copytree(src, dst)
+
+        src_machine = join(dirname(__file__), 'asv-machine.json')
+        dst_machine = join(cache_dir, 'asv-machine.json')
+        shutil.copyfile(src_machine, dst_machine)
+
+        # Convert to current file format
+        conf = config.Config.from_json({'results_dir': dst,
+                                        'repo': 'none',
+                                        'project': 'asv'})
+        run_asv_with_conf(conf, 'update', _machine_file=dst_machine)
+
+        return dst
